VFB-414: renamed + move file
