+++ src/components/Tables/AddressColumnFormatter.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { ClientsTableRow } from "@/app/clients/clientsTable/types";
 import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
 import ClientOutsideDeliveryAreaIcon from "@/components/Icons/ClientsOutsideDeliveryAreaIcon";
-import { formatNullPostcode } from "@/common/format";
+import { getDisplayPostcode } from "@/common/format";
 
 export const rowToAddressColumn = ({
     postcode,
     isDeliverable,
 }: ClientsTableRow["addressPostcode"] | ParcelsTableRow["addressPostcode"]): React.ReactElement => {
     const postcodeRow: React.ReactNode[] = [];
-    postcodeRow.push(formatNullPostcode(postcode));
+    postcodeRow.push(getDisplayPostcode(postcode));
     if (!isDeliverable && postcode !== "-") {
         postcodeRow.push(
             <span style={{ paddingLeft: "0.3rem" }}>
